task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,16 +7,16 @@
  */
 int main(void)
 {
-	int n1 = 1, n2 = 2;
-	int nextTerm;
+	long n1 = 1, n2 = 2;
+	long nextTerm;
 	int i;
 
-	printf("%d, %d, ", n1, n2);
+	printf("%ld, %ld, ", n1, n2);
 
 	for (i = 1; i <= 50; i++)
 	{
 		nextTerm = n1 + n2;
-		printf("%d, ", nextTerm);
+		printf("%ld, ", nextTerm);
 		n1 = n2;
 		n2 = nextTerm;
 	}
